chore: move ui and worker files into directories
