+++ proton-pass-web/src/lib.rs
@@ -1,20 +1,10 @@
 use wasm_bindgen::prelude::*;
 
-#[cfg(feature = "web_ui")]
-mod creditcard;
-#[cfg(feature = "web_ui")]
-mod login;
+mod common;
+
 #[cfg(feature = "web_ui")]
 mod ui;
 
-#[cfg(feature = "web_worker")]
-mod common;
-#[cfg(feature = "web_worker")]
-mod passkey;
-#[cfg(feature = "web_worker")]
-mod password;
-#[cfg(feature = "web_worker")]
-mod utils;
 #[cfg(feature = "web_worker")]
 mod worker;
