+++ proton-pass-web/src/worker/password.rs
